Create sockets before domain participant
